Moved all pairwise velocity functions into their own dedicated modules
