+++ halotools/mock_observables/pairwise_velocities/tests/test_los_pvd_vs_rp.py
@@ -4,7 +4,7 @@ import numpy as np
 from astropy.tests.helper import pytest
 from astropy.utils.misc import NumpyRNGContext
 
-from ..pairwise_velocity_stats import los_pvd_vs_rp
+from ..los_pvd_vs_rp import los_pvd_vs_rp
 
 from ...tests.cf_helpers import generate_locus_of_3d_points
